+++ solution/WebApplication/WebApplication/Controllers/SubjectAreaController.cs
@@ -4,9 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore;
 using WebApplication.Services;
-
+using WebApplication.Framework;
 using WebApplication.Models;
 
 namespace WebApplication.Controllers
@@ -16,7 +17,7 @@ namespace WebApplication.Controllers
         protected readonly AdsGoFastContext _context;
         
 
-        public SubjectAreaController(AdsGoFastContext context, SecurityAccessProvider securityAccessProvider) : base(securityAccessProvider)
+        public SubjectAreaController(AdsGoFastContext context, ISecurityAccessProvider securityAccessProvider, IEntityRoleProvider roleProvider) : base(securityAccessProvider, roleProvider)
         {
             Name = "SubjectArea";
             _context = context;
@@ -25,10 +26,12 @@ namespace WebApplication.Controllers
         // GET: SubjectArea
         public async Task<IActionResult> Index()
         {
-            return View(await _context.SubjectArea.ToListAsync());
+            var adsGoFastContext = _context.SubjectArea.Include(s => s.SubjectAreaForm);
+            return View(await adsGoFastContext.ToListAsync());
         }
 
         // GET: SubjectArea/Details/5
+        [ChecksUserAccess]
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -37,11 +40,13 @@ namespace WebApplication.Controllers
             }
 
             var subjectArea = await _context.SubjectArea
+                .Include(s => s.SubjectAreaForm)
                 .FirstOrDefaultAsync(m => m.SubjectAreaId == id);
             if (subjectArea == null)
-            {
                 return NotFound();
-            }
+            if (!await CanPerformCurrentActionOnRecord(subjectArea))
+                return new ForbidResult();
+
 
             return View(subjectArea);
         }
@@ -49,7 +54,10 @@ namespace WebApplication.Controllers
         // GET: SubjectArea/Create
         public IActionResult Create()
         {
-            return View();
+            ViewData["SubjectAreaFormId"] = new SelectList(_context.SubjectAreaForm.OrderBy(x=>x.SubjectAreaFormId), "SubjectAreaFormId", "SubjectAreaFormId");
+     SubjectArea subjectArea = new SubjectArea();
+            subjectArea.ActiveYn = true;
+            return View(subjectArea);
         }
 
         // POST: SubjectArea/Create
@@ -57,18 +65,25 @@ namespace WebApplication.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [ChecksUserAccess]
         public async Task<IActionResult> Create([Bind("SubjectAreaId,SubjectAreaName,ActiveYn,SubjectAreaFormId,DefaultTargetSchema,UpdatedBy")] SubjectArea subjectArea)
         {
             if (ModelState.IsValid)
             {
                 _context.Add(subjectArea);
+                if (!await CanPerformCurrentActionOnRecord(subjectArea))
+                {
+                    return new ForbidResult();
+                }
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(IndexDataTable));
             }
+        ViewData["SubjectAreaFormId"] = new SelectList(_context.SubjectAreaForm.OrderBy(x=>x.SubjectAreaFormId), "SubjectAreaFormId", "SubjectAreaFormId", subjectArea.SubjectAreaFormId);
             return View(subjectArea);
         }
 
         // GET: SubjectArea/Edit/5
+        [ChecksUserAccess()]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -78,9 +93,11 @@ namespace WebApplication.Controllers
 
             var subjectArea = await _context.SubjectArea.FindAsync(id);
             if (subjectArea == null)
-            {
                 return NotFound();
-            }
+
+            if (!await CanPerformCurrentActionOnRecord(subjectArea))
+                return new ForbidResult();
+        ViewData["SubjectAreaFormId"] = new SelectList(_context.SubjectAreaForm.OrderBy(x=>x.SubjectAreaFormId), "SubjectAreaFormId", "SubjectAreaFormId", subjectArea.SubjectAreaFormId);
             return View(subjectArea);
         }
 
@@ -89,6 +106,7 @@ namespace WebApplication.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [ChecksUserAccess]
         public async Task<IActionResult> Edit(int id, [Bind("SubjectAreaId,SubjectAreaName,ActiveYn,SubjectAreaFormId,DefaultTargetSchema,UpdatedBy")] SubjectArea subjectArea)
         {
             if (id != subjectArea.SubjectAreaId)
@@ -101,6 +119,10 @@ namespace WebApplication.Controllers
                 try
                 {
                     _context.Update(subjectArea);
+
+                    if (!await CanPerformCurrentActionOnRecord(subjectArea))
+                        return new ForbidResult();
+			
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -114,12 +136,14 @@ namespace WebApplication.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(IndexDataTable));
             }
+        ViewData["SubjectAreaFormId"] = new SelectList(_context.SubjectAreaForm.OrderBy(x=>x.SubjectAreaFormId), "SubjectAreaFormId", "SubjectAreaFormId", subjectArea.SubjectAreaFormId);
             return View(subjectArea);
         }
 
         // GET: SubjectArea/Delete/5
+        [ChecksUserAccess]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -128,24 +152,31 @@ namespace WebApplication.Controllers
             }
 
             var subjectArea = await _context.SubjectArea
+                .Include(s => s.SubjectAreaForm)
                 .FirstOrDefaultAsync(m => m.SubjectAreaId == id);
             if (subjectArea == null)
-            {
                 return NotFound();
-            }
+		
+            if (!await CanPerformCurrentActionOnRecord(subjectArea))
+                return new ForbidResult();
 
             return View(subjectArea);
         }
 
         // POST: SubjectArea/Delete/5
         [HttpPost, ActionName("Delete")]
+        [ChecksUserAccess()]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var subjectArea = await _context.SubjectArea.FindAsync(id);
+
+            if (!await CanPerformCurrentActionOnRecord(subjectArea))
+                return new ForbidResult();
+		
             _context.SubjectArea.Remove(subjectArea);
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction(nameof(IndexDataTable));
         }
 
         private bool SubjectAreaExists(int id)
